Move OnSurroundingTextChange validation logic to individual suggesters.
This text logic is only used by emoji and personal info, and some of the logic (e.g. checking whitespace at the end) is being doubled up. This separates the responsibility of analysing the text to the individual suggesters, and makes the responsibility of the main assistivesuggester clearer (it just passes on the event and coordinates the suggesters).

Changed the definition of TrySuggestWithSurroundingText to take int (A.Suggester::TrySuggest... uses int so there was an unnecessary conversion). Also put anchor pos in so that this logic can be done for each individual suggesters.

Edge case for when a new line exists after cursor:
This is the expected behaviour for each case (current behaviour before this cl's code changes)
- In Emoji Suggester: Does not suggest
- In Personal Info Suggester: Does suggest

Bug: b/225988020
Change-Id: I31a45aac953879aa1f63c69d046652823f095408
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3583879
Reviewed-by: Darren Shen <shend@chromium.org>
Reviewed-by: Jing Wang <jiwan@chromium.org>
Commit-Queue: jhtin <jhtin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#994512}