Remove cache of video transformation in WebMediaPlayerMS
There is a known issue that the video width/height of HTML video element
will give incorrect values under portrait orientation.

When getting the natural size of WebMediaPlayerMS, since it uses the
current frame size of the compositor and the cached video transformation
to calculate, sometimes the result will be desynced and cause the issue
mentioned above.

This CL removes the cached video transformation in WebMediaPlayerMS and
instead query the current frame from compositor and then extract the
information from the frame.

Bug: b:229175191
Test: blink_unittests, blink_tests
Change-Id: Ia07c75683e8321f784e4377b94d1385fa84cf32e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3586535
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Wei Lee <wtlee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#994526}