Move Mojo and IPC protobuf support to separate targets
Currently, any target depending on Mojo or IPC must also depend on
protobuf, but the only dependency is some helper traits headers. As a
result, we still need to build protobuf for NaCl.

Split these into separate targets. This removes the last protobuf
dependency in NaCl and removes the need to build it for NaCl at all. In
particular, the new version of protobuf does not build with NaCl's libc
(it expects an <endian.h>). Rather than patch that, remove the
dependency altogether.

Bug: 1294200
Change-Id: If4e448e681c94a5eed329c57ef08f4dd88be5437
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3594979
Reviewed-by: Joe Downing <joedow@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Commit-Queue: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/main@{#994534}