Pause socket managers for a page when it enters the BFCache.
This prevents updated network conditions from being IPCed to the frozen
page, which is a potential privacy issue.

This is done by having the P2PSocketDispatcherHost keep a mapping of
renderer frame to socket_manager, which allows the render frame host to
request that only socket managers for that particular frame be paused
when the frame enters the BFCache.

Bug: 1287029
Change-Id: I4e0e86d5ffe739ad8f218dbea70d982572ae04e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3412316
Reviewed-by: Harald Alvestrand <hta@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Commit-Queue: Darren Willis <djw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#994571}