[WebApk-Weblayer] Weblayer client for installing WebAPKs using the WebApkInstallCoordinator-service in Chrome.
This CL enables the WebApk installation via BottomSheet/BannerUI or
Ambient Badge in Weblayer. If a website fulfills PWA criteria and
site-engagement is reached, preferably a bottom sheet with an install
button is shown, alternatively a BannerUI or Ambient Badge. When the
user triggers the install the WebAPK-proto is built in WebLayer and
passed, together with a callback, to Chrome's WebApkInstallCoordinator-
service.
(CL: https://chromium-review.googlesource.com/c/chromium/src/+/3503251)
The callback is called by the service as soon as the installation
finished or failed.
The binary size increase is caused by newly implemented classes, the
additional dependency of Promise.java (which is used for triggering the
nested callback functions when connecting to the Chrome-service and
the forwarding of the callback result which was received from the
service) and a dependency to //components/signin which seems to be
the result of some internal optimization of the build, but the overall
increase through that is within limits.

Follow up CLs:
- Display a Toast in the failure case and when the installation starts.
- Only show the install-banner if a Chrome version that contains the
  service is installed on the device.
- UMA logging.
- BrowserTest for WebApkInstallScheduler.

Binary-Size: See commit description.
Bug: 1101999
Change-Id: I7cb9a48235873c607efd12c83332c0bfbdc23865
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3542366
Reviewed-by: Glenn Hartmann <hartmanng@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Peter Conn <peconn@chromium.org>
Commit-Queue: Susanne Westphal <swestphal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#994580}