[WebApk-Weblayer] Show install bottom sheet only if Chrome with WebApkInstallCoordinator service is available.
The install bottom sheet/install banner should only be displayed if the
installation is actually possible, so if Chrome with a version that
exposes the WebApkInstallCoordinator service is installed.

This CL adds the check if the service is available and only shows the
install option in that case.

Bug: 1101999
Change-Id: I22f1c0137bd0469f4aabf5939f7dab46796041c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3575380
Reviewed-by: Peter Conn <peconn@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Susanne Westphal <swestphal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#994597}