Add tests to latency info trace events data
The Interactions swimlane in DevTools Performance panel was broken due
to changes in the shape of the data of latency info trace events.
Because of the lack of tests to the data in these events, the breakage
in the panel wasn't noticed until more than 2 years after the first
breaking changes landed. This CL adds tests to the generation of data
so that we can catch future breaking changes before they cause a
regression in the frontend.

The tests added in this CL make sure that the data is dispatched in the
way (shape) expected by the implementation of
https://crrev.com/c/3583855 (which is based on the up-to-date shape of
the latency info trace events).

Bug: 1317621
Change-Id: I05ec040507bcbc5bcb2663dcabdc83dc9318cb87
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3594760
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#994672}