Fix flakiness of the AccessibilityEventsImmediateRefresh test on Linux
The purpose of this test is to verify there is no DCHECK failure during
a load-complete event -- an event which is normally fired before the
event recorder is logging events. Due to a to-be-determined timing
issue, sometimes a focus event is logged prior to test completion
whereas other times it is not. Because the exact events are irrelevant
for the purpose of this test, this commit ensures no events are dumped
for Linux.

Bug: 1316465
Change-Id: I171796046ad40c0883fa29eba7297d6bbaf36972
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3599413
Reviewed-by: Martin Robinson <mrobinson@igalia.com>
Commit-Queue: Joanmarie Diggs <jdiggs@igalia.com>
Cr-Commit-Position: refs/heads/main@{#994696}