Ensure core is disconnected prior to being destroyed
Some classes that observe CloudPolicyCore and have longer lifespan, rely
on OnCoreDisconnecting callbacks to unregister themselves prior to core
being destroyed. We ensure that this happens in the destructor. If the
core was never connected or disconnected prior to the being destroyed,
this is a no-op and no additional callbacks are invoked.

Bug: 1217542
Change-Id: I04719fccc97d5dfe0ffa79942f750eac3f8b45ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3593262
Auto-Submit: Sergiy Belozorov <sergiyb@chromium.org>
Reviewed-by: Fabio Tirelo <ftirelo@chromium.org>
Commit-Queue: Fabio Tirelo <ftirelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#994700}