[Android][PwdCheck] Remove assert crashing for multiply compromised pwds
The assert wanted to guard that compromised credentials are either
leaked or compromised (which is a correct precondition for the UI).

That works if they are *only* leaked or *only* phished but not if the
credential is compromised in multiple ways.
Leaked *and* phished could trigger this.
Leaked and weak or leaked and phished and weak also trigger that which
made this bug more prominent.

The params are renamed (and properly computed) in crrev.com/c/3599273.
Lucky for us, the assert only behaves like a DCHECK so no releases have
been impacted.

Bug: 1318328
Change-Id: I37052bb4fea57008afba56869a38e3db054060a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3599277
Commit-Queue: Friedrich Horschig <fhorschig@chromium.org>
Auto-Submit: Friedrich Horschig <fhorschig@chromium.org>
Reviewed-by: Viktor Semeniuk <vsemeniuk@google.com>
Commit-Queue: Viktor Semeniuk <vsemeniuk@google.com>
Cr-Commit-Position: refs/heads/main@{#994708}