Trunk prefix related local heuristics logic.
Assign correct phone number types depending on the presence of a country
code field in local heuristics.
Guarded behind |kAutofillEnableSupportForPhoneNumberTrunkTypes|.

Adds TrunkPrefixTypes unittest to verify the new behavior.

Removes max length constraints of phone part in <cc> - <phone> grammar.
Limiting the phone part of a common grammar like <cc> - <phone> to 14
(or 10, as the outdated comment suggests) is not ideal, because this
prevents classifying fields without max-length as such.
As a result the CountryAndCityAndPhoneNumberWithLongerMaxLength test was
removed.
This fixes the classification of one browser tests to correctly
recognize a <cc> - <phone> combination, instead of labeling it is
ADDRESS_HOME_COUNTRY and PHONE_HOME_WHOLE_NUMBER.
Ideally the length constraint on the <cc> part should be removed as
well, but this breaks code elsewhere.

Refactors `PhoneFieldTest` by eliminating common code between unittests.

As the final CL of trunk type implementation, marks them as
IsFillableFieldType.

Bug: 1311937
Change-Id: Icf53d6e13ac8527cdf4073396ef050623dc1c552
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3562942
Reviewed-by: Christoph Schwering <schwering@google.com>
Reviewed-by: Matthias Körber <koerber@google.com>
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#994725}