Fenced Frames: Fix ad size test that fails on screens with width 600
This test case expects (screen width)x90 to be coerced to (screen width)x100. But on Nexus 7, the screen width is 600, and the closer size 600x93 is available due to the 32/5 aspect ratio flexible size. We can fix this by using 99 rather than 90. (I verified it manually with the dev tools mobile screen simulator.)

Bug: 1317720, 1123606
Change-Id: If0352b2afe04c36a739984dfc510aed223484d2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3597054
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Commit-Queue: Garrett Tanzer <gtanzer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#994747}