Move STGButton to its own file in preparation for code split of BBV 1/4
We are going to split the functionality of BookmarkBarViews Tab Group
implementation out into its own view for testability and code clarity.
First we will start with what is already implemented in BBV, the
buttons for the Saved Tab Groups.

We also create a new AX node description for the saved tab group buttons.

Change-Id: I1bece0fe44ab6f79a0827afa9071ce1bc2b75332
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3561503
Reviewed-by: Darryl James <dljames@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: David Pennington <dpenning@chromium.org>
Cr-Commit-Position: refs/heads/main@{#994748}