Fixes the position of the delegate layer of shelf navigation widget.
Originally, the bounds of the delegate view layer of shelf navigation
widget wasn't calculated for RTL mode. This worked because the delegate
view and all its children were symmetric. With the new launcher nudge
with label view, the statement won't necessarily be true anymore, and
we want to put the delegate view and the buttons on the right corner of
the screen.
With this cl, we assume that the navigation widget delegate locates at
the corner of the screen, which is bottom left in LTR and bottom right
in RTL.

Bug: 1308733
Change-Id: I03a2a1f55a2220b31ee1e12e5be7fcdc5bcffb08
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3596909
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: Wen-Chien Wang <wcwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#994751}