[iOS] Error page refactoring, fix error page restoration
The LoadSimulatedRequest API was not handle by the restoration as
expected. It always creates a new navigation item when there is no
currently committed item, so if the first navigation after restoration
triggers an error page, it would create a new item instead of using the
existing one.

To avoid that, when it is a restoration, we use loadHTMLString instead
of loadSimulatedRequest.

See result video:
https://drive.google.com/file/d/1AOlZ-xzI-NkJhcMMlMjBQ6jFfs3UQLXt/view?usp=sharing

Change-Id: I25eaac2da0a8e8ad619febca178fc6ab86a507d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3599073
Reviewed-by: Ali Juma <ajuma@chromium.org>
Commit-Queue: Aliona Dangla <alionadangla@chromium.org>
Auto-Submit: Aliona Dangla <alionadangla@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#994753}