lacros: Fix crash due to unpinned ash-chrome app.
If the ash-chrome app is not available in the listed of pinned apps,
then ash will crash. It's not immediately clear how this can happen as
it is always pinned at startup. Regardless of the cause, the existing
logic should be robust to either lacros or ash as the pinned app.

Bug: 1316653
Change-Id: I50045d1b4162068bd627f918dcf5a65b43324c21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3597518
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#994755}