[Extensions] Ensure deleted/unreadable content script files are verified
ExtensionUserScriptLoader/LoadScriptContent logs warning when it
encounters a deleted / unreadable content script file, and skips
"content verification (CV)" for it. This CL propagates the info
to CV in those cases too (as it does for "regular" files), so
files that were tampered with can be detected.

This CL adds regression tests for crbug/1296310

Bug: 1296310
Change-Id: Iebcf9ddd40839cebf77dd479197086eaff0b49ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3587139
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Istiaque Ahmed <lazyboy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#994786}