[side search] Eliminate grey flash when opening Side Search
This CL eliminates the grey flash that can be seen when side search
is first opened.

This is done by setting the default background color of the
WebContents to be transparent and setting the hosting WebView's
background color to the toolbar background color.

Before:
https://screencast.googleplex.com/cast/NTQxMzQ3MzE1NzY0NDI4OHw1ZTRhOTEwMS1kOQ

After:
https://screencast.googleplex.com/cast/NjQ4MjY1MDQ3OTk4NDY0MHwwNDZhMmY5MS0wYg

Bug: 1315412
Change-Id: I86a8879ddae422a3c1355924af012bd41d2adf3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3598181
Reviewed-by: Yuheng Huang <yuhengh@chromium.org>
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#994790}