[infra] Add consistency checks to builder config generation.
The recipe code requires that the majority of the fields be equal on the
builder specs for the builders (not testers) that are mirrored by a try
builder. This adds the same consistency check to the builder config
generation so such errors are caught at config generation time instead
of when the recipe runs.

Bug: 1314422
Change-Id: If0b5e3d4aba6d57cc89fe4eba57833d81c8c8634
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3598311
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#994806}