Mac: don't send accessibility events on nodes that aren't in a window
Product justification: if the node isn't in a window, the change isn't
relevant to the user.

Technical justification: this has caused multiple hangs. See
https://crbug.com/1298082#c45 and https://crbug.com/1221262
I think this is probably an undocumented limitation of the
API (or at least the old API; see bug for details).


Bug: 1298082
Change-Id: Id5c81e56ef64740793c967afeadfc452c10e3526
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3595084
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Leonard Grey <lgrey@chromium.org>
Cr-Commit-Position: refs/heads/main@{#994812}