lacros: Refactor chromeosInfoPrivate.
This CL is a refactor with no intended behavior change.

This CL moves the business logic in info_private_api.h to a new class
ExtensionInfoPrivateAsh. It adds some glue code in the form of a new crosapi. This will also be used by Lacros to implemented chromeosInfoPrivate. The only new code is the glue code itself.

Bug: 1318109
Change-Id: Ie0a5e9e597db4055fcc303bcaa023788dde48e9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3596905
Reviewed-by: Jenny Zhang <jennyz@chromium.org>
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#994838}