Add a Separate View for Saved Tab Groups in the bookmarks bar.
Here a new view is added which serves as a container for the Saved
Tab Group Buttons. In this CL it only handles implementing the needed
functions from the SavedTabGroupModelObserver and some helper functions.

The Bookmarks Bar implementation of the SavedTabGroupButtons is then
removed and in its place the SavedTabGroupBar is added.

Change-Id: Ib2a4e71cba5fd407fcd558e60678221323f9964c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3561465
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: David Pennington <dpenning@chromium.org>
Cr-Commit-Position: refs/heads/main@{#994839}