Remove svg style element tree test from Android
This CL removes a broken unit test from the
WebContentsAccessibilityTreeTest suite on Android.

The test was recently changed and began to flake. This is because
there is no mechanism to handle <script> elements on the external
Android tree tests. All such tests were ignored when the suite was
first created, so we have removed this test to match.

AX-Relnotes: N/A
Bug: 1317065
Change-Id: I9e667e71f24abcf48eede12951c6f5bd8be7f2f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3601009
Reviewed-by: Abigail Klein <abigailbklein@google.com>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#994860}