cros: fix the crash of the cross fade animation
This CL fixs the crash of cross fade animation when using dual display.

The crash is mainly caused by
`FrameHeader::FrameAnimatorView::StartAnimation` that
recreates the window layer when setting the cross fade animation.

When setting the old layer animation, it will stops all the running
animations. If ExitAnimation of Overview is in the sequence, the
OverviewController will destroy the ExitAnimation and set the window
paint as active. Then, `FrameHeader::FrameAnimatorView::StartAnimation`
is triggered and the window's layer is recreated. In this case, the
precached new layer is no longer the window layer and is set up an
opacity animation.

Later, when the new layer sets opacity, the current opacity animation is
stopped and
`FrameHeader::FrameAnimatorView::OnImplicitAnimationCompleted` is
triggered which will destroy the precached new layer which causes
the crash.

To solve this, we should update the new layer after the old layer stops
animation.

The unittest simulates this case.

Test: WindowAnimationsTest.RecreateWhenSettingCrossFade
Bug: 1313977
Change-Id: Ifc22a616fe4790544f79846b5fee3220fec04fc9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3598633
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Commit-Queue: Xiaodan Zhu <zxdan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#994872}