[iOS] Number of lines is now respected for Answer suggestions
This CL implements custom number of lines in Answer suggestions for the
new omnibox popup, as in the previous version of the popup. Specifically
the detail text view now uses the suggested number of lines and ellipsis
for truncation, instead of just displaying one line with a truncating
gradient just like regular suggestions.

Fixed: 1316638
Change-Id: I6a89fd1e8c65145fb44b914cc01b9b80fd5dcf3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3593258
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#994874}