Switch document Picture in Picture (v2) to popup
Since we are including more "normal" browser features, like
permissions, in picture in picture windows, it makes sense to
normalize pip windows to be their own browser window type.

Previously, pip popup windows used a window that was created and
managed by a DocumentPictureInPictureWindowController.  This work
is now done by the browser, similar to how regular popup windows are
created.  In the future, the business logic will still reside in a
special pip controller (e.g., "close the pip window when the
opener window navigates away"), but this CL turns all of this off
to keep the CL more focused.

Since pip v2 is currently turned off behind a flag, this doesn't
break anything.

As part of this, this CL turns off some of the pip tests that don't
work anymore.  These will be re-enabled in a follow-up CL that
fixes the pip window controller.

Change-Id: I4aa3cbffc5e1cef91d1c99b0e7eaf0583732ed7e
Bug: 1317445
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3576675
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Fr <beaufort.francois@gmail.com>
Commit-Queue: Frank Liberato <liberato@chromium.org>
Cr-Commit-Position: refs/heads/main@{#994933}