personalization: Regenerate Google Photos photo placeholders on resize.
Because Google Photos photos are manually partitioned into rows based on
window size we need to explicitly regenerate placeholders when the
desired number of photos per row changes.

Note that the non-placeholder case is already WAI but we need to be
explicit with placeholders since the triggering resize event will not
result in a change to the partitioned `photosBySection_` when `photos_`
is null|undefined (aka the placeholder case).

Bug: b:229906960
Change-Id: I409c923aee27986a106bfd38bf17fc883a109a57
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3598051
Reviewed-by: Jeffrey Young <cowmoo@chromium.org>
Commit-Queue: David Black <dmblack@google.com>
Cr-Commit-Position: refs/heads/main@{#994957}